Make Syntax Much More Generic
The new syntax file is less specific about the format it expects the log
file to be in. It should work more-or-less correctly with log files from
various applications, programming languages, servers, etc. We'll
continue to improve it as we find things that don't look good.

In place of tests, we'll include a sample.log to see how it works. And
add a screenshot of this to the README.